email-sending: no moving items
It’s a lot of dance for the sole purpose of making the process
inspectable. This moving of items around adds its own complexity, so I’m
moving towards just reporting more to the console. It’ll be up to the
program execution env to decide what to do with that information, but I
will provide it.