Housekeeping: logs: better double-click UX
When readin the watch-app emails, the log record comes prettified like
this:

{
  "severity": "info",
  "message": "User registered",
  "data": {
    "module": "registration:initAccount",
    "plan": "minimal",
    "email": {
      "kind": "EmailAddress",
      "value": "gurdiga@gmail.com"
    },
    "hashedPassword": "[**masked**]",
    "creationTimestamp": "2022-11-02T16:23:37.992Z"
  }
}

Double-clicking the module name "registration:initAccount" selects
everything between the quotes, which is not what I want usually. I want
to be able to double click and select "initAccount" then copy-and-paste
it to find that function in the code.