Post mortem: loadFeed hashingSalt
DEFINITIONS

Feed hashingSalt is only used for computing email’s salted hash. Email
salted hash is used as its opaque unique ID.

SYNOPSIS

The bug was that when loading an existing feed’s data, the hashingSalt
was being randomly generated every time.

It didn’t cause any trouble actually, because feed.hashingSalt is used
to generate email hash, which is stored only on subscription, and from
then on is used directly, with no implication of feed.hashingSalt.

So, having feed.hashingSalt randomized on every feedLoad was not
affecting anything else for the resto of the email’s lifetime.