api: GET /feeds/:feedId: no checkSchedule
How to reasonably display the chec schedule?

Ana mentioned something along the lines of: Maybe don’t display at all?

When is it useful to change it?

I think I had this little discomfort when I was presenting the concept
of check schedule, as if the argument for having it felt a little
artificial.

I could go with a blanket statement like this: At the moment we check
hourly.

Having the schedule unreachable to the user, reduces the conceptual
complexity for the user.

DECISION: Remove it.